\name{interpretR-package}
\alias{interpretR-package}
\alias{interpretR}
\docType{package}
\title{
Partial Dependence Plots and Permutation-Based Performance Measures
}
\description{
Compute permutation-based performance meaures (for binary classification) and create partial dependence plots (cross-validated classification and regression models). Currently only binary classification and regression models estimated with the package \code{randomForest} are supported. Binary classification models estimated with \code{ada} are also supported.
}
\details{
\tabular{ll}{
Package: \tab interpretR\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2014-03-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Authors: Michel Ballings, and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}
\seealso{
\code{\link[interpretR:parDepPlot]{parDepPlot}}, \code{\link[interpretR:variableImportance]{variableImportance}}
}