% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpretRNews.R
\name{interpretRNews}
\alias{interpretRNews}
\title{Display the NEWS file}
\usage{
interpretRNews()
}
\description{
\code{interpretRNews} shows the NEWS file of the interpretR package.
}
\examples{
interpretRNews()
}
\author{
Authors: Michel Ballings and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}
\seealso{
\code{\link{parDepPlot}}
}

