\name{interpretR-package}
\alias{interpretR-package}
\alias{interpretR}
\docType{package}
\title{
Binary Classifier Interpretation functions: Partial Dependence Plots and Permutation-Based Performance Measures
}
\description{
Compute permutation-based performance meaures and create partial dependence plots for binary (cross-validated) classification models. Currently only binary classification models estimated with the packages \code{randomForest} and \code{ada} are supported.
}
\details{
\tabular{ll}{
Package: \tab interpretR\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2014-11-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Authors: Michel Ballings, and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}
\seealso{
\code{\link[interpretR:parDepPlot]{parDepPlot}}, \code{\link[interpretR:variableImportance]{variableImportance}}
}