\name{interpretR-package}
\alias{interpretR-package}
\alias{interpretR}
\docType{package}
\title{
Partial Dependence Plots and Permutation-Based Performance Measures
}
\description{
Compute permutation-based performance meaures (for binary classification) and create partial dependence plots (cross-validated classification and regression models). Currently only binary classification and regression models estimated with the package \code{randomForest} are supported. Binary classification models estimated with \code{ada} are also supported.
}
\author{
Authors: Michel Ballings, and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}
\seealso{
\code{\link[interpretR:parDepPlot]{parDepPlot}}, \code{\link[interpretR:variableImportance]{variableImportance}}
}